Renamed test file
